requirements.txt
requests==2.32.3
fastapi==0.115.8
httpx==0.28.1
uvicorn==0.34.0
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytz==2025.1
python-dotenv==1.0.1